replaced all size type to size_t. long long to intmax_t. size checks are now performed against SIZE_MAX >> 1

+++ lib/generics/include/vector.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stddef.h>
+#include <stdint.h>
 
 #include "defines.h"
 
@@ -10,15 +12,15 @@
 /* vector object */
 struct vector {
   // both size and capacity can never exceed LLONG_MAX
-  unsigned long long size;
-  unsigned long long capacity;
-  unsigned long long data_size;
+  size_t size;
+  size_t capacity;
+  size_t data_size;
   unsigned char *data;
 };
 
 /* initialize a vector object. returns struct vector * on success, NULL on
  * failure */
-struct vector *vector_init(unsigned long long data_size);
+struct vector *vector_init(size_t data_size);
 
 /* destroy a vector and all if it's undelying data. if (*destroy) isn't NULL
  * call it for every element in the underlying array. you should only pass in a
@@ -28,18 +30,18 @@ void vector_destroy(struct vector *vector, void (*destroy)(void *element));
 
 /* returns the number of elements in the vector. avoid acceessing vector::size
  * directly. use this method instead */
-unsigned long long vector_size(struct vector *vector);
+size_t vector_size(struct vector *vector);
 
 /* returns the number of elements you can fit in the vector. avoid acceessing
  * vector::capacity directly. use this method instead */
-unsigned long long vector_capacity(struct vector *vector);
+size_t vector_capacity(struct vector *vector);
 
 /* returns whether vector is emtpy or not. if vector is NULL - returns true */
 bool vector_empty(struct vector *vector);
 
 /* returns the element at position pos. any changes to the element will change
  * the stored element on the vector. returns NULL on failure */
-void *vector_at(struct vector *vector, unsigned long long pos);
+void *vector_at(struct vector *vector, size_t pos);
 
 /* sorts the vector and finds an element on the vector and returns it. returns
  * NULL on failure */
@@ -47,7 +49,7 @@ void *vector_find(struct vector *vector, const void *element, int (*cmpr)(const 
 
 /* reservse space for size elements. returns the new reserved space
  * (vector::capacity) */
-unsigned long long vector_reserve(struct vector *vector, unsigned long long size);
+size_t vector_reserve(struct vector *vector, size_t size);
 
 /* changes the size of the vector. if size < vector::size vector::size will
  * decrease to the size passed in. beware if the vector contains a pointers to
@@ -56,7 +58,7 @@ unsigned long long vector_reserve(struct vector *vector, unsigned long long size
  * followed by vector_resize. if size >= vector::size && size <
  * vector::capacity, vector::size will be set to size and number of NULL values
  * will be pushed into the vector. returns the new vector::size */
-unsigned long long vector_resize(struct vector *vector, unsigned long long size);
+size_t vector_resize(struct vector *vector, size_t size);
 
 /* push a new element into the vector. returns true on success, false otherwise
  */
@@ -70,21 +72,21 @@ void *vector_pop(struct vector *vector);
 
 /* remove the element at position pos. returns the removed element (which has to
  * be free'd). NULL on failure */
-void *vector_remove_at(struct vector *vector, unsigned long long pos);
+void *vector_remove_at(struct vector *vector, size_t pos);
 
 /* replaces an element on the vector at position pos. returns the
  * replaced element on success as heap allocated element (has to be free'd), or
  * NULL on failure */
-void *vector_replace(struct vector *vector, const void *element, unsigned long long pos);
+void *vector_replace(struct vector *vector, const void *element, size_t pos);
 
 /* shrink the underlying array to fit exactly vector::size elements. returns the
  * new capacity */
-unsigned long long vector_shrink(struct vector *vector);
+size_t vector_shrink(struct vector *vector);
 
 /* finds and returns the index of the first occurence of an element on the
  * vector. returns its position on success, or N_EXISTS if no such element
  * found */
-long long vector_index_of(struct vector *vector, const void *element, int (*cmpr)(const void *, const void *));
+intmax_t vector_index_of(struct vector *vector, const void *element, int (*cmpr)(const void *, const void *));
 
 /* sort the vector. the compr function should returns an int bigger than 0 if
  * the first element if bigger the second, 0 if both elements are equals or an
